♻️ refactor: 업로드 메서드 수정, 폴더 이름 변경 #9
1. getUrl 메서드 제거, 업로드 시 최종 파일 주소 반환하도록 변경
2. 파일명: error -> exception
