+++ src/main/java/com/example/teamB/global/upload/exception/UploadException.java
@@ -1,4 +1,4 @@
-package com.example.teamB.global.upload.error;
+package com.example.teamB.global.upload.exception;
 
 import com.example.teamB.global.apiPayload.exception.CustomException;
