+++ src/main/java/com/example/teamB/global/upload/handler/UploadHandler.java
@@ -6,6 +6,5 @@ import org.springframework.web.multipart.MultipartFile;
  * 저장 방식 변경을 대비한 인터페이스
  */
 public interface UploadHandler {
-    void upload(MultipartFile file, String storeImageName);
-    String getUrl(String fileName);
+    String upload(MultipartFile file);
 }
